Example: Using Timer in strategies - Code review fixes
