+++ examples/backtest/example_001_load_bars_from_custom_csv/6EH4.XCME_1min_bars.csv
