+++ tests/test_data/xcme/README.md
@@ -1,22 +1,23 @@
-# File: `6EH4.GLBX_1min_bars_20240101_20240131.csv.zip`
+# File: `6EH4_1min_bars_20240101_20240131.csv.zip`
 
     - Instrument: 6E
     - Expiration: H4 (March 2024)
-    - Exchange:   CME Globex
+    - Exchange:   XCME (MIC code)
     - Period      2024-01-01 --> 2024-01-31 (UTC timestamped, no contract rollover occurs in this period)
     - Bar type:   1-minute bars
 
 
 # Zipped format
 
-We used zipped data, because they are 9x smaller and can be DIRECTLY read by [pandas](https://pandas.pydata.org/)
+We used zipped data, because they are 9x smaller than original CSV file and can be DIRECTLY read by [pandas](https://pandas.pydata.org/)
 using code like this:
 
 ```python
 import pandas as pd
 
+
 df = pd.read_csv(
-    "6EH4.GLBX_1min_bars_20240101_20240131.csv.zip",  # update path as needed
+    "6EH4_1min_bars_20240101_20240131.csv.zip",  # update path as needed
     sep=";",
     decimal=".",
     header=0,
